UserspaceEmulator: Split large file into logical pieces
