+++ Userland/DevTools/UserspaceEmulator/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(SOURCES
     Emulator.cpp
+    Emulator_syscalls.cpp
     MallocTracer.cpp
     MmapRegion.cpp
     Range.cpp
